Add: Better sound controller errors and debug Also fixed small issue, with volume. In previous version volume was fixed int, so almost any portion after dividing by 100 would by 0 which caused to have volume set to only max or 0.

+++ Assets/Scripts/Utilities/Settings/SoundSettings.cs
@@ -13,7 +13,6 @@ namespace Utilities.Settings
 
         public UniverseSoundNode.SoundType SettingSoundType => settingSoundType;
         public int SetVolume(int volume) => currentVolume = volume;
-        // ReSharper disable once PossibleLossOfFraction
-        public float CurrentVolume => currentVolume/100;
+        public float CurrentVolume => (float)currentVolume/100;
     }
 }
